fix router, bug with layout, added next/font instead of google font link
