+++ app/layout.tsx
@@ -1,24 +1,36 @@
-'use client';
-
-import { ThemeProvider } from '@mui/material';
-import { theme } from './styles/theme';
+import { Metadata } from 'next';
+// import { ThemeProvider } from '@mui/material';
+// import { theme } from './styles/theme';
 import Navbar from './components/features/Navbar';
 import './globals.css';
+import ProvidersWrapper from './ProvidersWrapper';
+
+export const metadata: Metadata = {
+  title: 'Days of Planty',
+  description: 'Plant diary',
+  keywords: ['Next.js', 'React', 'TypeScript', 'MongoDB', 'Material UI'],
+  authors: [
+    {
+      name: 'Haeju eom',
+      url: 'https://github.com/kaulfield23',
+    },
+  ],
+  creator: 'Haeju Eom',
+  publisher: 'Haeju Eom',
+};
 
 export default function RootLayout({
-  // Layouts must accept a children prop.
-  // This will be populated with nested layouts or pages
   children,
 }: {
   children: React.ReactNode;
 }) {
   return (
     <html lang="en">
-      <body suppressHydrationWarning={true}>
-        <ThemeProvider theme={theme}>
+      <body>
+        <ProvidersWrapper>
           <Navbar />
           {children}
-        </ThemeProvider>
+        </ProvidersWrapper>
       </body>
     </html>
   );
