+++ package.json
@@ -13,22 +13,17 @@
     "@emotion/styled": "^11.10.5",
     "@mui/icons-material": "^5.11.11",
     "@mui/material": "^5.11.9",
-    "@next/font": "13.1.6",
     "@types/mongoose": "^5.11.97",
     "@types/node": "18.13.0",
     "@types/react": "18.0.28",
     "@types/react-dom": "18.0.11",
     "eslint": "8.34.0",
     "eslint-config-next": "^13.4.9",
-    "formidable": "^3.5.0",
     "mongodb": "^5.6.0",
     "mongoose": "^7.3.1",
     "next": "^13.4.9",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "typescript": "4.9.5"
-  },
-  "devDependencies": {
-    "@types/formidable": "^3.4.0"
   }
 }
